+++ lib/CMakeLists.txt
@@ -42,6 +42,7 @@ add_subdirectory(Migrator)
 add_subdirectory(Option)
 add_subdirectory(Parse)
 add_subdirectory(PrintAsClang)
+add_subdirectory(Refactoring)
 add_subdirectory(RemoteAST)
 add_subdirectory(Sema)
 add_subdirectory(Serialization)
