+++ lib/IDE/CMakeLists.txt
@@ -28,7 +28,6 @@ add_swift_host_library(swiftIDE STATIC
   KeyPathCompletion.cpp
   ModuleInterfacePrinting.cpp
   PostfixCompletion.cpp
-  Refactoring.cpp
   REPLCodeCompletion.cpp
   SourceEntityWalker.cpp
   SwiftSourceDocInfo.cpp
@@ -42,7 +41,6 @@ target_link_libraries(swiftIDE PRIVATE
   swiftAST
   swiftClangImporter
   swiftDriver
-  swiftIndex
   swiftParse
   swiftSema)
